// src/app/models/system/policy/reasons.ts
export class ReasonsModel {
    constructor() {

    }
    company_code: string = "";
    reason_id: string = "";
    reason_code: string = "";
    reason_name_th: string = "";
    reason_name_en: string = "";
    reason_group: string="";
    created_by: string = "";
    created_date: string = "";
    modified_by: string = "";
    modified_date: string = "";
    flag: boolean = false;
}